requirements.txt
django 
django-modeltranslation
whitenoise